app/requirements.txt
uvicorn
fastapi
pydantic

singleton_package
json-advanced
python-dotenv
debugpy
ipython

aiohttp
aiofiles
aiocache
usso

beanie
fastapi-mongo-base

metisai
langdetect
pillow
apscheduler